Prepare core integration: fritzboxvpn library, runtime_data, PyPI workflow, and HA.io docs draft.
Extract API access into the fritzboxvpn PyPI package, align the integration with Home Assistant core patterns, and add publish/documentation assets for the upstream merge path.
