+++ .github/workflows/ci.yml
@@ -7,6 +7,7 @@ on:
       - main
     paths: &ci-paths
       - "custom_components/**"
+      - "fritzboxvpn/**"
       - "tests/**"
       - "scripts/**"
       - "hacs.json"
@@ -45,7 +46,7 @@ jobs:
 
       - uses: astral-sh/ruff-action@v3
         with:
-          src: custom_components tests
+          src: custom_components tests fritzboxvpn/fritzboxvpn
 
   pytest:
     name: pytest
