+++ .github/workflows/publish-pypi.yml
@@ -1,0 +1,56 @@
+# Publish the fritzboxvpn PyPI package (required for Home Assistant Core).
+#
+# Setup (once): docs/pypi-publish.md
+
+name: Publish fritzboxvpn to PyPI
+
+on:
+  push:
+    tags:
+      - "v*"
+  workflow_dispatch:
+    inputs:
+      confirm:
+        description: 'Type "publish" to upload the current fritzboxvpn version to PyPI'
+        required: true
+
+concurrency:
+  group: publish-pypi-${{ github.ref }}
+  cancel-in-progress: false
+
+permissions:
+  contents: read
+
+jobs:
+  publish:
+    name: PyPI
+    if: github.event_name != 'workflow_dispatch' || github.event.inputs.confirm == 'publish'
+    runs-on: ubuntu-latest
+    timeout-minutes: 15
+    environment: pypi
+    permissions:
+      contents: read
+      id-token: write
+    steps:
+      - uses: actions/checkout@v6
+
+      - uses: actions/setup-python@v6
+        with:
+          python-version: "3.13"
+
+      - name: Install build tools
+        run: python -m pip install --upgrade pip build twine
+
+      - name: Build wheel and sdist
+        working-directory: fritzboxvpn
+        run: python -m build
+
+      - name: Verify package metadata
+        working-directory: fritzboxvpn
+        run: twine check dist/*
+
+      - name: Publish to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v2
+        with:
+          packages-dir: fritzboxvpn/dist
+          print-hash: true
