+++ custom_components/fritzbox_vpn/binary_sensor.py
@@ -2,37 +2,30 @@
 
 from typing import Any
 
+from fritzboxvpn import API_KEY_CONNECTED
 from homeassistant.components.binary_sensor import (
     BinarySensorDeviceClass,
     BinarySensorEntity,
 )
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .const import (
-    API_KEY_CONNECTED,
-    DATA_COORDINATOR,
-    DATA_KNOWN_UIDS_BINARY_SENSOR,
-    DATA_LOCK_ADD_ENTITIES_BINARY_SENSOR,
-    DOMAIN,
-    UNIQUE_ID_SUFFIX_CONNECTED,
-)
+from .const import UNIQUE_ID_SUFFIX_CONNECTED
 from .coordinator import FritzBoxVPNCoordinator
 from .entity import FritzBoxVPNEntity, setup_vpn_platform
+from .models import FritzboxVpnConfigEntry
 
 
 async def async_setup_entry(
     hass: HomeAssistant,
-    entry: ConfigEntry,
+    entry: FritzboxVpnConfigEntry,
     async_add_entities: AddEntitiesCallback,
 ) -> None:
     """Set up FritzBox VPN binary sensor entities."""
-    coordinator: FritzBoxVPNCoordinator = hass.data[DOMAIN][entry.entry_id][
-        DATA_COORDINATOR
-    ]
 
-    def _create_entities(uids: set[str]) -> list[FritzBoxVPNConnectedBinarySensor]:
+    def _create_entities(
+        coordinator: FritzBoxVPNCoordinator, uids: set[str]
+    ) -> list[FritzBoxVPNConnectedBinarySensor]:
         if not coordinator.data:
             return []
         return [
@@ -42,12 +35,9 @@ async def async_setup_entry(
         ]
 
     await setup_vpn_platform(
-        hass,
         entry,
         async_add_entities,
-        platform_label="binary_sensor",
-        known_uids_key=DATA_KNOWN_UIDS_BINARY_SENSOR,
-        lock_key=DATA_LOCK_ADD_ENTITIES_BINARY_SENSOR,
+        platform="binary_sensor",
         create_entities=_create_entities,
     )
 
@@ -58,7 +48,7 @@ class FritzBoxVPNConnectedBinarySensor(FritzBoxVPNEntity, BinarySensorEntity):
     def __init__(
         self,
         coordinator: FritzBoxVPNCoordinator,
-        entry: ConfigEntry,
+        entry: FritzboxVpnConfigEntry,
         connection_uid: str,
         connection_data: dict[str, Any],
     ) -> None:
