+++ custom_components/fritzbox_vpn/const.py
@@ -3,7 +3,11 @@
 from collections.abc import Mapping
 from typing import Any
 
+from fritzboxvpn import const as fritzboxvpn_const
 from homeassistant.const import CONF_HOST, CONF_PASSWORD
+
+DEFAULT_NAME_UNKNOWN = fritzboxvpn_const.DEFAULT_NAME_UNKNOWN
+NAME_FRITZBOX = fritzboxvpn_const.NAME_FRITZBOX
 
 DOMAIN = "fritzbox_vpn"
 CONF_UPDATE_INTERVAL = "update_interval"
@@ -14,24 +18,6 @@ DEFAULT_UPDATE_INTERVAL = 30
 UPDATE_INTERVAL_MIN = 5
 UPDATE_INTERVAL_MAX = 3600
 RETRY_AFTER_SECONDS = 300
-DEFAULT_TIMEOUT = 10
-DEFAULT_PROTOCOL = "https"
-VERIFICATION_DELAY = 1.5
-
-API_LOGIN = "/login_sid.lua"
-API_DATA = "/data.lua"
-API_VPN_CONNECTION = "/api/v0/generic/vpn/connection/{uid}"
-
-# API request/response keys (Fritz!Box data.lua / boxConnections)
-API_PAGE_SHAREWIREGUARD = "shareWireguard"
-API_KEY_DATA = "data"
-API_KEY_INIT = "init"
-API_KEY_BOX_CONNECTIONS = "boxConnections"
-API_KEY_UID = "uid"
-API_KEY_ACTIVE = "active"
-API_KEY_ACTIVATED = "activated"
-API_KEY_CONNECTED = "connected"
-API_KEY_NAME = "name"
 
 ATTR_UID = "uid"
 ATTR_VPN_UID = "vpn_uid"
@@ -49,53 +35,6 @@ UNIQUE_ID_SUFFIXES = (
     UNIQUE_ID_SUFFIX_SWITCH,
     UNIQUE_ID_SUFFIX_UID,
 )
-ERROR_MSG_INVALID_SID = "Invalid SID"
-ERROR_MSG_INVALID_SID_403 = "Invalid SID (HTTP 403)"
-ERROR_MSG_INVALID_SID_HTML = "Invalid SID (HTML response)"
-ERROR_MSG_LOGIN_FAILED_SID = (
-    "Login failed: Invalid SID. This can be caused by: "
-    "(1) Incorrect username or password, or "
-    "(2) TR-064 not being enabled. "
-    "Please check your credentials first, then verify that TR-064 (Permit access for apps) "
-    "is enabled in the {name_fritzbox} under "
-    "Home Network > Network > Network settings > Access Settings in the Home Network. "
-    "Note: UPnP is only needed for automatic discovery via SSDP, not for API access."
-)
-PROTOCOL_HTTP = "http"
-PROTOCOL_HTTPS = "https"
-PROTOCOLS_ALLOWED = (PROTOCOL_HTTP, PROTOCOL_HTTPS)
-CONTENT_TYPE_JSON = "json"
-
-HTTP_STATUS_OK = 200
-HTTP_STATUS_FORBIDDEN = 403
-HTTPS_FALLBACK_STATUS_CODES = (400, 404, 502, 503)
-
-LOGIN_TAG_CHALLENGE = "Challenge"
-LOGIN_TAG_SID = "SID"
-LOGIN_TAG_BLOCKTIME = "BlockTime"
-LOGIN_FORM_USERNAME = "username"
-LOGIN_FORM_RESPONSE = "response"
-INVALID_SID_VALUE = "0000000000000000"
-
-LOG_LABEL_ACTIVATED = "activated"
-LOG_LABEL_DEACTIVATED = "deactivated"
-ACTIVE_STATE_STRINGS_TRUE = ("1", "true", "yes", "on")
-HEADER_VALUE_APPLICATION_JSON = "application/json"
-AUTH_HEADER_PREFIX = "AVM-SID "
-
-DATA_COORDINATOR = "coordinator"
-DATA_FRITZ_SESSION = "fritz_session"
-DATA_KNOWN_UIDS_SWITCH = "known_uids_switch"
-DATA_KNOWN_UIDS_SENSOR = "known_uids_sensor"
-DATA_KNOWN_UIDS_BINARY_SENSOR = "known_uids_binary_sensor"
-DATA_KNOWN_UIDS_KEYS = (
-    DATA_KNOWN_UIDS_SWITCH,
-    DATA_KNOWN_UIDS_SENSOR,
-    DATA_KNOWN_UIDS_BINARY_SENSOR,
-)
-DATA_LOCK_ADD_ENTITIES_SWITCH = "lock_add_entities_switch"
-DATA_LOCK_ADD_ENTITIES_SENSOR = "lock_add_entities_sensor"
-DATA_LOCK_ADD_ENTITIES_BINARY_SENSOR = "lock_add_entities_binary_sensor"
 
 OPTIONS_ACTION_CONFIGURE = "configure"
 OPTIONS_ACTION_CLEANUP = "cleanup"
@@ -112,10 +51,8 @@ LOG_MSG_VPN_CONNECTIONS_REMOVED_HINT = (
 )
 
 MANUFACTURER_AVM = "AVM"
-NAME_FRITZBOX = "Fritz!Box"
 MODEL_FRITZBOX = "Fritz!Box"
 MODEL_WIREGUARD_VPN = "WireGuard VPN"
-DEFAULT_NAME_UNKNOWN = "Unknown"
 INTEGRATION_TITLE = "Fritz!Box VPN"
 NOTIFICATION_TITLE_AUTH_ERROR = "Fritz!Box VPN: Authentifizierungsfehler"
