+++ custom_components/fritzbox_vpn/diagnostics.py
@@ -2,18 +2,20 @@
 
 from typing import Any
 
+from fritzboxvpn import API_KEY_ACTIVE, API_KEY_CONNECTED, API_KEY_NAME
 from homeassistant.components.diagnostics import async_redact_data
 from homeassistant.const import CONF_PASSWORD, CONF_USERNAME
 from homeassistant.core import HomeAssistant
 
-from .const import CONF_UPDATE_INTERVAL, DATA_COORDINATOR, DOMAIN, host_from_config
+from .const import CONF_UPDATE_INTERVAL, host_from_config
 from .coordinator import normalize_update_interval
+from .models import FritzboxVpnConfigEntry, runtime_from_entry
 
 TO_REDACT = {CONF_USERNAME, CONF_PASSWORD, "password", "pass", "user", "username"}
 
 
 async def async_get_config_entry_diagnostics(
-    hass: HomeAssistant, entry: Any
+    hass: HomeAssistant, entry: FritzboxVpnConfigEntry
 ) -> dict[str, Any]:
     """Return diagnostics for a config entry (no credentials)."""
     host = host_from_config(entry.data)
@@ -25,23 +27,22 @@ async def async_get_config_entry_diagnostics(
     last_update_success: bool | None = None
     vpn_connections: list[dict[str, Any]] = []
 
-    if DOMAIN in hass.data:
-        store = hass.data[DOMAIN].get(entry.entry_id, {})
-        coordinator = store.get(DATA_COORDINATOR)
-        if coordinator is not None:
-            last_update_success = coordinator.last_update_success
-            if coordinator.data:
-                for uid, conn in coordinator.data.items():
-                    if not isinstance(conn, dict):
-                        continue
-                    vpn_connections.append(
-                        {
-                            "connection_uid": uid,
-                            "name": conn.get("name"),
-                            "active": conn.get("active"),
-                            "connected": conn.get("connected"),
-                        }
-                    )
+    runtime = runtime_from_entry(entry)
+    if runtime is not None:
+        coordinator = runtime.coordinator
+        last_update_success = coordinator.last_update_success
+        if coordinator.data:
+            for uid, conn in coordinator.data.items():
+                if not isinstance(conn, dict):
+                    continue
+                vpn_connections.append(
+                    {
+                        "connection_uid": uid,
+                        API_KEY_NAME: conn.get(API_KEY_NAME),
+                        API_KEY_ACTIVE: conn.get(API_KEY_ACTIVE),
+                        API_KEY_CONNECTED: conn.get(API_KEY_CONNECTED),
+                    }
+                )
 
     return {
         "entry": async_redact_data(entry.as_dict(), TO_REDACT),
