+++ custom_components/fritzbox_vpn/manifest.json
@@ -8,6 +8,13 @@
   "integration_type": "device",
   "iot_class": "local_polling",
   "issue_tracker": "https://github.com/rosch100/fritzbox-vpn/issues",
-  "requirements": ["aiohttp>=3.8.0"],
+  "loggers": ["fritzboxvpn"],
+  "quality_scale": "gold",
+  "requirements": ["fritzboxvpn==1.0.0"],
+  "ssdp": [
+    {
+      "st": "urn:schemas-upnp-org:device:fritzbox:1"
+    }
+  ],
   "version": "1.2.0"
 }
