+++ fritzboxvpn/README.md
@@ -1,0 +1,5 @@
+# fritzboxvpn
+
+Async Python library for the AVM Fritz!Box Web UI API (WireGuard VPN connections).
+
+Used by the Home Assistant `fritzbox_vpn` integration.
