+++ fritzboxvpn/pyproject.toml
@@ -1,0 +1,31 @@
+[project]
+name = "fritzboxvpn"
+version = "1.0.0"
+description = "Async Python library for AVM Fritz!Box WireGuard VPN Web API"
+readme = "README.md"
+requires-python = ">=3.11"
+license = "Apache-2.0"
+authors = [{ name = "rosch100" }]
+dependencies = ["aiohttp>=3.8.0"]
+
+[project.urls]
+Homepage = "https://github.com/rosch100/fritzbox-vpn"
+Repository = "https://github.com/rosch100/fritzbox-vpn"
+Issues = "https://github.com/rosch100/fritzbox-vpn/issues"
+Documentation = "https://github.com/rosch100/fritzbox-vpn/tree/main/fritzboxvpn"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["fritzboxvpn"]
+
+[tool.ruff]
+target-version = "py311"
+line-length = 88
+src = ["fritzboxvpn", "tests"]
+
+[tool.ruff.lint]
+select = ["E", "F", "I", "UP", "B", "SIM"]
+ignore = ["E501"]
