+++ pyproject.toml
@@ -8,7 +8,7 @@ requires-python = ">=3.11"
 [tool.ruff]
 target-version = "py311"
 line-length = 88
-src = ["custom_components", "tests"]
+src = ["custom_components", "tests", "fritzboxvpn/fritzboxvpn"]
 
 [tool.ruff.lint]
 select = [
