+++ tests/test_binary_sensor_dynamic.py
@@ -2,11 +2,6 @@
 
 import pytest
 from custom_components.fritzbox_vpn import binary_sensor
-from custom_components.fritzbox_vpn.const import (
-    DATA_COORDINATOR,
-    DATA_KNOWN_UIDS_BINARY_SENSOR,
-    DOMAIN,
-)
 from homeassistant.core import HomeAssistant
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
@@ -17,8 +12,8 @@ from tests.fixtures import MOCK_VPN_CONNECTIONS
 async def test_binary_sensor_adds_on_coordinator_update(
     hass: HomeAssistant, coordinator_with_data, mock_config_entry: MockConfigEntry
 ) -> None:
-    """Coordinator listener registers new binary_sensor entities."""
-    coordinator = hass.data[DOMAIN][mock_config_entry.entry_id][DATA_COORDINATOR]
+    """Coordinator listener registers new binary sensor entities."""
+    coordinator = mock_config_entry.runtime_data.coordinator
     captured_listener = None
     original_add_listener = coordinator.async_add_listener
 
@@ -40,11 +35,11 @@ async def test_binary_sensor_adds_on_coordinator_update(
         "conn-new": {
             "uid": "wg-new",
             "name": "New",
-            "active": True,
-            "connected": True,
+            "active": False,
+            "connected": False,
         },
     }
-    hass.data[DOMAIN][mock_config_entry.entry_id][DATA_KNOWN_UIDS_BINARY_SENSOR] = set(
+    mock_config_entry.runtime_data.known_uids_binary_sensor = set(
         MOCK_VPN_CONNECTIONS.keys()
     )
     captured_listener()
